feat(DMR-01): create home page of application
Create a home page to show which transactions we made
and list it for the user. We cards to show inputs,
outputs and total of value.

Create modal to use when I need to add a new transaction in a
application. Everything about it was used Radix to have accessibility
and support to basic functions.

We using Radio and Dialog component of Radix in this commit.

Resolves: DMR-01-01, DMS-01-02