feat(DMR-02): create context and configure axios
Create the first Context API to manage Transactions in a application and
all initial configuration to use JSON Server as a Back-End If you have
but in this moment, we need to simulate as a Back-End to prepare a
communication between Front-End and Back-End.

Install and configure things like React Hook Forms, create an example
hook and how can I use formatter when you need a format date and price.

Install and configure axios swiching of fetch, because it can make the
code more organized and lint to apply specific formats can be help to
organize code in general.

Resolves: DMR-02-01, DMR-02-02, DMR-02-03